+++ cultivar-discovery/src/main/java/com/readytalk/cultivar/discovery/test/package-info.java
@@ -1,0 +1,5 @@
+/**
+ * Utilities for doing rapid prototyping with services that will eventually use ZK, but are not presently.
+ */
+@javax.annotation.ParametersAreNonnullByDefault
+package com.readytalk.cultivar.discovery.test;