+++ CHANGES.md
@@ -38,3 +38,4 @@ Changelog
 * Upgrading Gradle to 2.3.
 * Fixed lack of `CuratorService`s preventing demo code from working (Issue #55).
 * Adding property override to NodeContainer (Issue #54).
+* Adding tools to facilitate easier testing and rapid prototyping (Issue #60).