+++ CHANGES.md
@@ -39,3 +39,4 @@ Changelog
 * Fixed lack of `CuratorService`s preventing demo code from working (Issue #55).
 * Adding property override to NodeContainer (Issue #54).
 * Adding tools to facilitate easier testing and rapid prototyping (Issue #60).
+* Added the NodeCacheWrapper (Issue #62).